+++ CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   and scalar values are auto-wrapped into single-element arrays when the
   WDL type expects `Array[T]`
   ([#820](https://github.com/stjude-rust-labs/sprocket/pull/820)).
+* `sprocket format --newline-style` and `format.newline_style` config option to control the style of newlines in
+  `sprocket format` output ([#795](https://github.com/stjude-rust-labs/sprocket/pull/795)).
 
 ### Changed